feat: header 구현 및 레이아웃 고정
