+++ src/app/forest/layout.tsx
@@ -1,3 +1,4 @@
+import Header from "@/components/header";
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -12,6 +13,7 @@ export default function Layout({
 }>) {
   return (
     <>
+      <Header />
       {children}
       {/* <BottomNavigation /> */}
     </>
