+++ src/app/forest/page.tsx
@@ -1,3 +1,3 @@
 export default function ForestPage() {
-  return <main className="h-screen overflow-auto pb-16">Forest Page</main>;
+  return <main>Forest Page</main>;
 }
