+++ src/app/fundraisings/page.tsx
@@ -1,5 +1,3 @@
 export default function FundraisingsPage() {
-  return (
-    <main className="h-screen overflow-auto pb-16">Fundraisings Page</main>
-  );
+  return <main>Fundraisings Page</main>;
 }
