+++ src/app/home/page.tsx
@@ -1,3 +1,3 @@
 export default function Home() {
-  return <main className="h-screen overflow-auto pb-16">Landing Page</main>;
+  return <main>Landing Page</main>;
 }
